split : IndexTime 코드 분리
