+++ Sources/SwiftCueSheet/CueStruct/Index/IndexTime.swift
@@ -1,12 +1,12 @@
 //
-//  Index.swift
-//  CueSheet
+//  File.swift
+//  
 //
-//  Created by Aoikazto on 2020/06/07.
-//  Copyright © 2020 Aoikazto. All rights reserved.
+//  Created by Aoikazto on 2020/10/29.
 //
 
 import Foundation
+
 
 extension String {
     func leftPadding(toLength: Int, withPad character: Character) -> String {
@@ -78,35 +78,6 @@ public struct CSIndexTime : CustomStringConvertible {
     
     public init(min:Int, sec:Int, frame:Int){
         self.frames = frame + (sec * CSIndexTime.framePerSecond) + (min * CSIndexTime.framePerSecond * 60)
-    }
-    
-}
-
-public struct CSIndex {
-    public init(num:UInt8, time:CSIndexTime) {
-        _indexNum = num
-        _indexTime = time
-    }
-    
-    private var _indexNum:UInt8 = 0
-    private var _indexTime:CSIndexTime
-    
-    public private(set) var indexNum: UInt8 {
-        get {
-            return _indexNum
-        }
-        set {
-            _indexNum = newValue
-        }
-    }
-    
-    public var indexTime:CSIndexTime {
-        get {
-            return _indexTime
-        }
-        set {
-            _indexTime = newValue
-        }
     }
     
 }
